feat: Add support to `HttpMethod.all` method.
If any route is registered via this method, then that route will be returned no matter via which method the route is invoked.

Also while looking-up under certain method, if the route is not-found then it will automatically fallback to lookup under `HttpMethod.all` route. So first the priority will be given to the method passed in the `lookup` function.

Signed-off-by: tejHackerDev <tejasimha222@gmail.com>